Change all sensor to new structure
